+++ saleor/checkout/complete_checkout.py
@@ -14,6 +14,7 @@ from ..checkout import calculations
 from ..checkout.error_codes import CheckoutErrorCode
 from ..core.exceptions import InsufficientStock
 from ..core.taxes import TaxError, zero_taxed_money
+from ..core.tracing import traced_atomic_transaction
 from ..core.utils.url import validate_storefront_url
 from ..discount import DiscountInfo, DiscountValueType, OrderDiscountType
 from ..discount.models import NotApplicable
@@ -32,7 +33,7 @@ from ..order.models import Order, OrderLine
 from ..order.notifications import send_order_confirmation
 from ..payment import PaymentError, gateway
 from ..payment.models import Payment, Transaction
-from ..payment.utils import store_customer_id
+from ..payment.utils import fetch_customer_id, store_customer_id
 from ..product.models import ProductTranslation, ProductVariantTranslation
 from ..warehouse.availability import check_stock_quantity_bulk
 from ..warehouse.management import allocate_stocks
@@ -325,7 +326,7 @@ def _prepare_order_data(
     return order_data
 
 
-@transaction.atomic
+@traced_atomic_transaction()
 def _create_order(
     *,
     checkout_info: "CheckoutInfo",
@@ -512,6 +513,7 @@ def _get_order_data(
 
 def _process_payment(
     payment: Payment,
+    customer_id: Optional[str],
     store_source: bool,
     payment_data: Optional[dict],
     order_data: dict,
@@ -532,6 +534,7 @@ def _process_payment(
                 payment=payment,
                 token=payment.token,
                 manager=manager,
+                customer_id=customer_id,
                 store_source=store_source,
                 additional_data=payment_data,
                 channel_slug=channel_slug,
@@ -582,8 +585,13 @@ def complete_checkout(
         gateway.payment_refund_or_void(payment, manager, channel_slug=channel_slug)
         raise exc
 
+    customer_id = None
+    if store_source and payment:
+        customer_id = fetch_customer_id(user=user, gateway=payment.gateway)
+
     txn = _process_payment(
         payment=payment,  # type: ignore
+        customer_id=customer_id,
         store_source=store_source,
         payment_data=payment_data,
         order_data=order_data,